Drop debcargo generated autopkgtests
These don't work since we're not even building a binary -dev package.
The packaging was originally created via dh-cargo and when that needed
too many overrides it was decided to not package it via debcargo-conf
and instead do a traditional separate git packaging repo. The
debian/tests/control file was carried over, but it will not work so drop
it.